feat : reply구현 및 이미지 업로드 구현중

+++ src/components/community/detail/comment/CommunityDetailComment.tsx
@@ -1,54 +1,61 @@
 'use client';
 
 import styled from '@emotion/styled';
-import { Text } from '@components/ui/Text';
 import timeFormat from '@utils/timeFormat';
-import { useRouter } from 'next/navigation';
 import Image from 'next/image';
 import DefaultProfileThumbnail from 'public/images/profile/profile.png';
-import { ColorToken } from 'styles/Color';
-import { useState } from 'react';
-import CommunityDetailCommentReplyList from './CommunityDetailCommentReplyList';
-import CommunityDetailCommentReplyInput from './CommunityDetailCommentReplyInput';
-import QuerySuspenseErrorBoundary from '@components/hoc/QuerySuspenseErrorBoundary';
-import { Button } from '@components/ui/Button';
-import { useMutation, useQueryClient } from '@tanstack/react-query';
-import deleteCommunityDetailComment from 'src/api/community/comment/deleteCommunityDetailComment';
 import DropDown from '@components/ui/DropDown';
 import CommentDropDown from './CommentDropDown';
+import deleteCommunityDetailComment from 'src/api/community/comment/deleteCommunityDetailComment';
 import EditIcon from '@components/icons/common/Edit.icon';
 import DeleteIcon from '@components/icons/common/Delete.icon';
-import { API_GET_COMMUNITY_DETAIL_COMMENT_KEY } from 'src/api/community/comment/getCommunityDetailComment';
-import { CommunitySearchParams } from 'types/searchParams/community';
+import CommunityDetailCommentInput from './CommunityDetailCommentInput';
+import { Text } from '@components/ui/Text';
+import { useState } from 'react';
+import { Typo } from 'styles/Typography';
+import { useRecoilState } from 'recoil';
+import { isCommentDropDownRecoil } from 'src/recoil-states/commentState';
+import { API_COMMUNITY_DETAIL_COMMENT_KEY } from 'src/api/community/comment/getCommunityDetailComment';
+import { API_COMMUNITY_DETAIL_COMMENT_REPLY_KEY } from 'src/api/community/comment/reply/getCommunityDetailCommentReply';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 export interface CommunityDetailCommentProps {
     item: any;
-    searchParams: CommunitySearchParams;
     params: {
-        id: string;
+        postId: string;
+        commentId?: string;
+        replyId?: string;
     };
+    type?: 'reply';
+    dropdownmenu?: any;
+    isEdit?: boolean;
 }
 
 const Container = styled.div`
     display: flex;
-    flex-direction: column;
+    flex-direction: row;
     width: 100%;
     align-items: start;
+    gap: 16px;
     margin-bottom: 24px;
 `;
+
 const CommentContainer = styled.div`
     display: flex;
-    justify-content: space-between;
-    align-items: center;
+    flex-direction: column;
     width: 100%;
 `;
-const ReplyContainer = styled.div`
-    position: relative;
+const InputContainer = styled.div`
     display: flex;
     width: 100%;
-    flex-direction: column;
+    margin-top: 24px;
+`;
+
+const CommentDetailContainer = styled.div`
+    display: flex;
+    justify-content: space-between;
     align-items: start;
-    padding-left: 52px;
+    width: 100%;
 `;
 
 const ProfileImage = styled(Image)`
@@ -75,68 +82,52 @@ const UserMeta = styled.div`
     flex-direction: row;
     gap: 6px;
 `;
-
-const ActionButtons = styled.div`
+const Time = styled.div`
     display: flex;
-    flex-direction: row;
-    gap: 16px;
-`;
-const ReplyButton = styled.div`
-    margin-top: 16px;
-    position: relative;
-    left: -16px;
+    min-width: 86px;
 `;
 
-const ReplyInputButton = styled.div`
-    span {
-        cursor: pointer;
-        &:hover {
-            text-decoration: underline;
-        }
-    }
+const CommentImage = styled(Image)`
+    object-fit: cover;
+    border-radius: 4px;
 `;
 
-const CommunityDetailComment = ({ item, params, searchParams }: CommunityDetailCommentProps) => {
-    const { id: postId } = params;
-    const { id, comment, username = 'user', timestamp, subcollectionCount: ReplyCount } = item;
-    const { push } = useRouter();
+const CommunityDetailComment = ({ item, params }: CommunityDetailCommentProps) => {
+    const { postId, commentId } = params;
+    const { id, content, username = 'user', timestamp, image } = item;
     const [isReplyInput, setIsReplyInput] = useState(false);
-    const [isReply, setIsReply] = useState(false);
-
+    const [openId, setOpenId] = useRecoilState(isCommentDropDownRecoil);
+    const [isEdit, setIsEdit] = useState(false);
     const queryCache = useQueryClient();
 
     const { mutate: deleteMutate } = useMutation({
         mutationFn: deleteCommunityDetailComment,
     });
 
-    const { mutate: editMutate } = useMutation({
-        mutationFn: deleteCommunityDetailComment,
-    });
-
     const deleteHandler = () => {
+        const userConfirm = window.confirm('정말 삭제하시겠습니까?');
+        if (!userConfirm) {
+            return;
+        }
+        const onSuccess = async () => {
+            await queryCache.invalidateQueries({
+                queryKey: [API_COMMUNITY_DETAIL_COMMENT_KEY, { postId }],
+            });
+            await queryCache.invalidateQueries({
+                queryKey: [API_COMMUNITY_DETAIL_COMMENT_REPLY_KEY, { postId, commentId }],
+            });
+        };
         deleteMutate(
-            { id },
+            { id, params },
             {
-                onSuccess: async () => {
-                    await queryCache.invalidateQueries({
-                        queryKey: [API_GET_COMMUNITY_DETAIL_COMMENT_KEY, params, searchParams],
-                    });
-                },
+                onSuccess: onSuccess,
             },
         );
     };
 
     const editHandler = () => {
-        editMutate(
-            { id },
-            {
-                onSuccess: async () => {
-                    await queryCache.invalidateQueries({
-                        queryKey: [],
-                    });
-                },
-            },
-        );
+        setIsEdit(true);
+        setOpenId(null);
     };
 
     const dropdownmenu = [
@@ -150,61 +141,74 @@ const CommunityDetailComment = ({ item, params, searchParams }: CommunityDetailC
             id: 'edit',
             label: '편집',
             icon: <EditIcon />,
-            onClick: () => {},
+            onClick: editHandler,
         },
     ];
 
+    const dropDownOnClickHandler = () => {
+        setOpenId((prevId) => (prevId === id ? null : id));
+    };
+
     return (
         <Container>
-            <CommentContainer>
-                <ProfileContainer>
-                    <ProfileImage alt="profile_image" src={DefaultProfileThumbnail} />
-                    <UserDetails>
-                        <Text text={username} size={14} weight={600} />
-                        <Text text={comment} size={15} />
-                        <UserMeta>
-                            <Text
-                                text={timeFormat(timestamp, 'comment')}
-                                size={13}
-                                color={ColorToken.grey}
+            <ProfileImage alt="profile_image" src={DefaultProfileThumbnail} />
+            {isEdit ? (
+                <CommunityDetailCommentInput
+                    id={id}
+                    params={params}
+                    defaultValues={{ content, image }}
+                    closeInput={() => setIsEdit(false)}
+                />
+            ) : (
+                <CommentContainer>
+                    <CommentDetailContainer>
+                        <ProfileContainer>
+                            <UserDetails>
+                                <Text text={username} typo={Typo.Body.Body2Bold} />
+                                {image &&
+                                    image.map(({ id, src }: any) => (
+                                        <CommentImage
+                                            key={id}
+                                            src={src}
+                                            alt="communitydetailimage"
+                                            width={200}
+                                            height={200}
+                                        />
+                                    ))}
+                                <Text text={content} />
+                                <UserMeta>
+                                    <Time>
+                                        <Text
+                                            text={timeFormat(timestamp, 'comment')}
+                                            typo={Typo.Body.Body3Regular}
+                                        />
+                                    </Time>
+                                    <Text
+                                        text="답글쓰기"
+                                        typo={Typo.Body.Body3Regular}
+                                        underLine
+                                        onClick={() => setIsReplyInput((e) => !e)}
+                                    />
+                                </UserMeta>
+                            </UserDetails>
+                        </ProfileContainer>
+                        <DropDown
+                            isOpen={openId === id}
+                            onClick={dropDownOnClickHandler}
+                            component={<CommentDropDown data={dropdownmenu} />}
+                        />
+                    </CommentDetailContainer>
+                    {isReplyInput && (
+                        <InputContainer>
+                            <CommunityDetailCommentInput
+                                id={id}
+                                params={params}
+                                closeInput={() => setIsReplyInput(false)}
                             />
-                            <ReplyInputButton>
-                                <Text
-                                    text="답글쓰기"
-                                    size={13}
-                                    color={ColorToken.grey}
-                                    onClick={() => {
-                                        setIsReplyInput((e) => !e);
-                                    }}
-                                />
-                            </ReplyInputButton>
-                        </UserMeta>
-                    </UserDetails>
-                </ProfileContainer>
-                <DropDown component={<CommentDropDown data={dropdownmenu} />} />
-            </CommentContainer>
-            <ReplyContainer>
-                {isReplyInput && (
-                    <CommunityDetailCommentReplyInput postId={postId} commentId={id} />
-                )}
-                {ReplyCount && (
-                    <>
-                        {isReply ? (
-                            <QuerySuspenseErrorBoundary>
-                                <CommunityDetailCommentReplyList postId={postId} commentId={id} />
-                            </QuerySuspenseErrorBoundary>
-                        ) : (
-                            <ReplyButton>
-                                <Button
-                                    text={`답글 ${ReplyCount}개`}
-                                    buttonType="reply"
-                                    onClick={() => setIsReply((e) => !e)}
-                                />
-                            </ReplyButton>
-                        )}
-                    </>
-                )}
-            </ReplyContainer>
+                        </InputContainer>
+                    )}
+                </CommentContainer>
+            )}
         </Container>
     );
 };
